Apply lib/patches/libh2o_wincompat_feature_gate_02.patch

+++ h2o/deps/picotls/deps/cifra/src/bitops.h
@@ -299,7 +299,7 @@ static inline void copy_bytes_unaligned(uint8_t *out, const uint8_t *in, size_t 
 static inline uint32_t count_trailing_zeroes(uint32_t x)
 {
 #ifdef _WINDOWS
-  uint32_t r = 0;
+  unsigned long r = 0;
   _BitScanReverse(&r, x);
   return (31 - r);
 #else
